Separate out output related functions
